refactor(27975): reorganise hooks for topic filters and tags
