+++ hivemq-edge/src/frontend/src/modules/Device/components/DeviceTagList.tsx
@@ -10,7 +10,7 @@ import ErrorMessage from '@/components/ErrorMessage.tsx'
 import { PLCTag } from '@/components/MQTT/EntityTag.tsx'
 import ArrayItemDrawer from '@/components/rjsf/SplitArrayEditor/components/ArrayItemDrawer.tsx'
 import { formatTagDataPoint } from '@/modules/Device/utils/tags.utils.ts'
-import { useTagManager } from '@/modules/Mappings/hooks/useTagManager.ts'
+import { useTagManager } from '@/modules/Device/hooks/useTagManager.ts'
 
 interface DeviceTagListProps {
   adapter: Adapter
