+++ hivemq-edge/src/frontend/src/modules/Device/hooks/useTagManager.ts
@@ -31,9 +31,7 @@ export const useTagManager = (adapterId: string) => {
 
     const { $schema: sc, ...rest } = tagSchema?.configSchema as RJSFSchema
     // TODO[28249] Handle manually until backend fixed
-    const { properties, required } = rest
-
-    const requiredProtocol = [...(required || []), 'protocolId']
+    const { properties } = rest
 
     const safeSchema = {
       ...rest,
@@ -44,8 +42,9 @@ export const useTagManager = (adapterId: string) => {
           default: protocol?.id,
         },
       },
-      required: requiredProtocol,
     }
+
+    console.log('XXXXXXXXXX safeSchema', safeSchema)
 
     return {
       // $schema: 'https://json-schema.org/draft/2020-12/schema',
@@ -110,7 +109,7 @@ export const useTagManager = (adapterId: string) => {
     )
   }
 
-  const context: ManagerContextType = {
+  const context: ManagerContextType<DomainTagList> = {
     schema: tagListSchema,
     uiSchema: {
       'ui:submitButtonOptions': {
