+++ hivemq-edge/src/frontend/src/modules/TopicFilters/hooks/useTopicFilterManager.spec.ts
@@ -3,7 +3,7 @@ import { renderHook, waitFor } from '@testing-library/react'
 
 import { server } from '@/__test-utils__/msw/mockServer.ts'
 import { SimpleWrapper as wrapper } from '@/__test-utils__/hooks/SimpleWrapper.tsx'
-import { useTopicFilterOperations } from '@/api/hooks/useTopicFilters/useTopicFilterOperations.ts'
+import { useTopicFilterManager } from '@/modules/TopicFilters/hooks/useTopicFilterManager.ts'
 import { handlers } from '@/api/hooks/useTopicFilters/__handlers__'
 
 import '@/config/i18n.config.ts'
@@ -14,7 +14,7 @@ describe('useTopicFilterManager', () => {
   })
 
   it('should return the manager', async () => {
-    const { result } = renderHook(() => useTopicFilterOperations(), { wrapper })
+    const { result } = renderHook(() => useTopicFilterManager(), { wrapper })
     expect(result.current.isLoading).toBeTruthy()
 
     await waitFor(() => {
